v0.77 fixed Add Task form

+++ component/admin/views/listtasks/tmpl/default.php
@@ -29,19 +29,28 @@ $jinput = JFactory::getApplication()->input;
 // **** Read data if a task was submitted... **** 
 $action = $jinput->get('action', '', 'STRING');
 $task_id = $jinput->get('task_id', 0, 'INT');
-$task_description = $jinput->get('task_description', '', 'STRING');
-$task_context = $jinput->get('context', '', 'STRING');
+$task_title = $jinput->get('task_title', '', 'STRING');
 
 if(!empty($action)) {
-  if(strcmp($action, "delete") == 0) {
+ 
+  if(strcmp($action, "delete") == 0 && $task_id > 0) {
     BaanabusHelper::deleteTask($db, $task_id);
   }
 
 }
-else if(!empty($task_description)) {
+else if(!empty($task_title) && empty($task_id)) {
+
+  // Create a new task from submitted strings & save to DB
+
   $new_task = new stdClass();
-  $new_task->task_description = $task_description; 
+  $new_task->task_title = $task_title; 
  
+  $task_description = $jinput->get('task_description', '', 'STRING');
+  if(!empty($task_description)) {
+    $new_task->task_description = $task_description; 
+  }
+
+  $task_context = $jinput->get('context', '', 'STRING');
   if(!empty($task_context)) {
     $new_task->context = $task_context; 
   }
@@ -65,41 +74,42 @@ echo "<table>";
 ?>
 
 <tr>
- <th> ID </th>
  <th style="width:200px; text-align:left;"> Title </th>
  <th style="width:200px; text-align:left;"> Description </th>
- <th style="width:100px; text-align:left;"> Doing it for... </th>
- <th width="200px"> Context </th>
+ <th style="width:200px; text-align:left;"> Doing it for... </th>
+ <th style="width:150px; text-align:left;"> Context </th>
+ <th style="width:100px; text-align:left;"> &nbsp; </th>
 </tr>
 
 <?php
 
 foreach ($tasks as $task) {
 
-echo "<tr>";
-echo "<td>" . $task->task_title . "</td> ";
-echo "<td>" . $task->task_description . "</td> ";
+  echo "<tr>";
+  echo "<td>" . $task->task_title . "</td> ";
+  echo "<td>" . $task->task_description . "</td> ";
 
-$person = function getPerson($db, $task->person_id);
+  $person = BaanabusHelper::getPerson($db, $task->person_id);
+  echo "<td>" . $person->name . "</td> ";
 
-echo "<td>" . $person->name . "</td> ";
-echo "<td>" . $task->context . "</td> ";
+  echo "<td>" . $task->context . "</td> ";
 
 
-?>
+  // then the edit button... 
+  ?>
+  <td>
+    <form action="index.php?option=com_baanabus&view=listtasks"  method="post" >
+    <input type="hidden" name="task_id" value="<?php echo $task->task_id; ?>"  />
+    <input type="hidden" name="action" value="delete"  />  
+    <input value="Delete" type="submit" class="btn btn-warning" >
+    </form>
+  </td>
+  <?php
 
-<td>
-  <form action="index.php?option=com_baanabus&view=listtasks"  method="post" >
-  <input type="hidden" name="task_id" value="<?php echo $task->task_id; ?>"  />
-  <input type="hidden" name="action" value="delete"  />  
-  <input value="Delete" type="submit" class="btn btn-warning" >
-  </form>
-</td>
+  echo "</tr>";
 
-</tr>
+} // end foreach(Task)
 
-<?php
-}
 echo "</table>";
 
 ?>
@@ -111,11 +121,11 @@ echo "</table>";
   <form action="index.php?option=com_baanabus&view=addtask" method="post" id="newTaskForm" name="newTaskForm">
 
   <br/>
-  <div class="label">Description: </div> 
-  <input type="text" class="form-control span6" name="task_description" value=""><br/>
+  <div class="label">Title: </div> 
+  <input type="text" class="form-control span6" name="task_title" ><br/>
 
   <div class="label">Description: </div> 
-  <input type="text" class="form-control span6" name="task_description" value=""><br/>
+  <input type="text" class="form-control span6" name="task_description" ><br/>
 
   <div class="label"> Context:</div> 
   <input type="text" class="form-control span6" name="context" value="home"> <br/>
